change: split box and product tests in separate files
